Introduce notion of repository, checkout and cookbook hashes.
The difference is this: if the files are located in the same place,
they have the same repository hash.  (imperfect at present because we
don't do URI normalization) If the files are located in the same place
and point to the same checkout, they have the same checkout hash.
There is an important difference between pointing to a branch and
pointing to a revision number, which is reflected.  Cookbook hashes
combine a checkout hash with a position in the checkout.

Add a cookbook hash.