Prevent recursing into cookbooks.
This essentially requires a state machine that
understands when it has just started to examine
a directory (and thus should check to see if it is
a cookbook, which is defined to be 'has a metadata.json
file').  The cleanest way to do that is to fetch
the first cookbook preemptively, and that required
the checkout scraper to be able to insert its setup
after the filesystem scraper had configured all the
variables it was going to use but before fetching
the first cookbook had happened.